pyproject.toml
[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[project]
name = "epistasis-v2"
version = "1.1.1"
description = "High-performance Python library for fitting high-order epistatic interactions in genotype-phenotype maps."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Unlicense" }
authors = [{ name = "Luis Perez", email = "lperezmo@users.noreply.github.com" }]
keywords = [
    "epistasis",
    "genotype-phenotype",
    "genetics",
    "genomics",
    "bioinformatics",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Rust",
    "License :: Public Domain",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "numpy>=1.23",
    "pandas>=2.0",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "lmfit>=1.2",
    "emcee>=3.1",
    "matplotlib>=3.7",
    "gpmap-v2>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/lperezmo/epistasis-v2"
Repository = "https://github.com/lperezmo/epistasis-v2"
Issues = "https://github.com/lperezmo/epistasis-v2/issues"

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-benchmark>=4",
    "pytest-cov>=5",
    "hypothesis>=6",
    "ruff>=0.7",
    "mypy>=1.11",
    "maturin>=1.7",
    "pandas-stubs>=2.3.3.260113",
]
examples = [
    "streamlit>=1.51",
    "plotly>=5.24",
]

[tool.maturin]
module-name = "epistasis._core"
python-source = "python"
manifest-path = "crates/epistasis-core/Cargo.toml"
features = ["pyo3/extension-module"]

[tool.semantic_release]
version_toml = [
    "pyproject.toml:project.version",
    "Cargo.toml:workspace.package.version",
]
version_variables = ["python/epistasis/_version.py:__version__"]
tag_format = "v{version}"
commit_message = "chore(release): v{version}"
# Release job does not build artifacts. Wheel + sdist builds happen in
# downstream matrix jobs via PyO3/maturin-action where Rust is available.
# Putting a build step here would fail: python-semantic-release runs in a
# Docker image with no Rust toolchain.
build_command = ""
# Force first release to go straight to 1.0.0 instead of climbing from
# 0.0.1. With this set, versions below 1.0.0 are treated as disallowed.
allow_zero_version = false

[tool.semantic_release.branches.main]
match = "main"
prerelease = false

[tool.semantic_release.commit_parser_options]
allowed_tags = [
    "feat",
    "fix",
    "perf",
    "refactor",
    "style",
    "docs",
    "test",
    "build",
    "ci",
    "chore",
    "revert",
]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["_legacy", "references"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
exclude = ["_legacy", "references", "tests"]
packages = ["epistasis"]
mypy_path = "python"

[[tool.mypy.overrides]]
module = [
    "gpmap",
    "gpmap.*",
    "sklearn",
    "sklearn.*",
    "lmfit",
    "lmfit.*",
    "emcee",
    "emcee.*",
    "scipy",
    "scipy.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --strict-markers"
